+++ witp/src/main/java/dev/efnilite/ip/hook/HoloHook.java
@@ -1,7 +1,7 @@
 package dev.efnilite.ip.hook;
 
 import dev.efnilite.ip.IP;
-import dev.efnilite.ip.api.Mode;
+import dev.efnilite.ip.mode.Mode;
 import dev.efnilite.ip.leaderboard.Leaderboard;
 import dev.efnilite.ip.leaderboard.Score;
 import me.filoghost.holographicdisplays.api.HolographicDisplaysAPI;