+++ witp/src/main/java/dev/efnilite/ip/mode/DefaultMode.java
@@ -1,10 +1,10 @@
 package dev.efnilite.ip.mode;
 
-import dev.efnilite.ip.api.Mode;
 import dev.efnilite.ip.config.Locales;
 import dev.efnilite.ip.config.Option;
 import dev.efnilite.ip.generator.ParkourGenerator;
 import dev.efnilite.ip.leaderboard.Leaderboard;
+import dev.efnilite.ip.menu.community.SingleLeaderboardMenu;
 import dev.efnilite.ip.player.ParkourPlayer;
 import dev.efnilite.ip.player.ParkourUser;
 import dev.efnilite.ip.session.Session;
@@ -12,21 +12,24 @@ import dev.efnilite.ip.util.Util;
 import dev.efnilite.vilib.inventory.item.Item;
 import org.bukkit.entity.Player;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 /**
  * The default parkour mode
  */
 public class DefaultMode implements Mode {
 
-    private final Leaderboard leaderboard = new Leaderboard(getName());
+    private final Leaderboard leaderboard = new Leaderboard(getName(), SingleLeaderboardMenu.Sort.SCORE);
 
     @Override
-    public @NotNull String getName() {
+    @NotNull
+    public String getName() {
         return "default";
     }
 
     @Override
-    public @NotNull Item getItem(String locale) {
+    @Nullable
+    public Item getItem(String locale) {
         return Locales.getItem(locale, "play.single.default");
     }
 
@@ -53,15 +56,5 @@ public class DefaultMode implements Mode {
                 .addPlayers(ParkourUser.register(player))
                 .generator(ParkourGenerator::new)
                 .complete();
-    }
-
-    @Override
-    public void click(Player player) {
-        create(player);
-    }
-
-    @Override
-    public boolean isVisible() {
-        return true;
     }
 }