+++ witp/src/main/java/dev/efnilite/ip/player/data/PreviousData.java
@@ -2,7 +2,6 @@ package dev.efnilite.ip.player.data;
 
 import dev.efnilite.ip.IP;
 import dev.efnilite.ip.config.Option;
-import dev.efnilite.ip.player.ParkourPlayer;
 import dev.efnilite.ip.reward.RewardString;
 import org.bukkit.GameMode;
 import org.bukkit.Location;
@@ -24,17 +23,13 @@ public class PreviousData {
     private Double maxHealth;
     private InventoryData inventoryData;
 
-    private final boolean allowFlight;
-    private final boolean flying;
-    private final boolean collidable;
-
     private final int hunger;
+    private final boolean flying;
+    private final boolean allowFlight;
     private final Player player;
     private final GameMode gamemode;
     private final Location location;
-
     private final Collection<PotionEffect> effects;
-
     public List<RewardString> rewardsLeaveList = new ArrayList<>();
 
     public PreviousData(@NotNull Player player) {
@@ -43,28 +38,23 @@ public class PreviousData {
         gamemode = player.getGameMode();
         location = player.getLocation();
         hunger = player.getFoodLevel();
-
         allowFlight = player.getAllowFlight();
         flying = player.getAllowFlight();
-        collidable = player.isCollidable();
-
-        if (Option.INVENTORY_HANDLING) {
-            this.inventoryData = new InventoryData(player);
-            this.inventoryData.saveInventory();
-            if (Option.INVENTORY_SAVING) {
-                this.inventoryData.saveFile();
-            }
-        }
-
         effects = player.getActivePotionEffects();
+
         for (PotionEffect effect : effects) {
             player.removePotionEffect(effect.getType());
         }
 
+        if (Option.INVENTORY_HANDLING) {
+            inventoryData = new InventoryData(player);
+            inventoryData.save(Option.INVENTORY_SAVING);
+        }
+
         // health handling after removing effects and inventory to avoid them affecting it
         if (Option.HEALTH_HANDLING) {
-            this.health = player.getHealth();
-            this.maxHealth = player.getAttribute(Attribute.GENERIC_MAX_HEALTH).getValue();
+            health = player.getHealth();
+            maxHealth = player.getAttribute(Attribute.GENERIC_MAX_HEALTH).getValue();
 
             player.getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(20);
             player.setHealth(maxHealth);
@@ -74,21 +64,14 @@ public class PreviousData {
     public void apply(boolean teleportBack) {
         try {
             if (teleportBack) {
-                if (Option.GO_BACK) {
-                    player.teleport(Option.GO_BACK_LOC);
-                } else {
-                    player.teleport(location);
-                }
+                player.teleport(Option.GO_BACK ? Option.GO_BACK_LOC : location);
             }
 
             player.setFoodLevel(hunger);
             player.setGameMode(gamemode);
-
             player.setAllowFlight(allowFlight);
             player.setFlying(flying);
-            player.setCollidable(collidable);
 
-            // -= Attributes =-
             if (maxHealth != null) {
                 player.getAttribute(Attribute.GENERIC_MAX_HEALTH).setBaseValue(maxHealth);
             }
@@ -96,27 +79,18 @@ public class PreviousData {
                 player.setHealth(health);
             }
 
-            // -= Potions =-
             for (PotionEffect effect : player.getActivePotionEffects()) {
                 player.removePotionEffect(effect.getType());
             }
             for (PotionEffect effect : effects) {
                 player.addPotionEffect(effect);
             }
-        } catch (Throwable ex) {// not optimal but there isn't another way
-            IP.logging().stack("Error while recovering stats of " + player.getName(), ex);
+        } catch (Exception ex) {// not optimal but there isn't another way
+            IP.logging().stack("Error while recovering stats of %s".formatted(player.getName()), ex);
         }
-        if (inventoryData != null) {
-            inventoryData.apply(false);
-        }
-    }
 
-    /**
-     * Applies the rewards stored in the leave list
-     *
-     * @param player The player to apply these rewards to
-     */
-    public void giveRewards(@NotNull ParkourPlayer player) {
-        rewardsLeaveList.forEach(s -> s.execute(player));
+        if (inventoryData != null) {
+            inventoryData.apply();
+        }
     }
 }